p4::ast

Struct Control

pub struct Control {
    pub name: String,
    pub variables: Vec<Variable>,
    pub constants: Vec<Constant>,
    pub type_parameters: Vec<String>,
    pub parameters: Vec<ControlParameter>,
    pub actions: Vec<Action>,
    pub tables: Vec<Table>,
    pub apply: StatementBlock,
}

Fields
name: Stringvariables: Vec<Variable>constants: Vec<Constant>type_parameters: Vec<String>parameters: Vec<ControlParameter>actions: Vec<Action>tables: Vec<Table>apply: StatementBlock
Implementations

impl Control

pub fn new(name: String) -> Self

pub fn get_parameter(&self, name: &str) -> Option<&ControlParameter>

pub fn get_action(&self, name: &str) -> Option<&Action>

pub fn get_table(&self, name: &str) -> Option<&Table>

pub fn tables<'a>( &'a self, ast: &'a AST, ) -> Vec<(Vec<(String, &'a Control)>, &'a Table)>

Return all the tables in this control block, recursively expanding local control block variables and including their tables. In the returned vector, the table in the second element of the tuple belongs to the control in the first element.

pub fn is_type_parameter(&self, name: &str) -> bool

pub fn names(&self) -> HashMap<String, NameInfo>

pub fn accept<V: Visitor>(&self, v: &V)

pub fn accept_mut<V: VisitorMut>(&self, v: &mut V)

pub fn mut_accept<V: MutVisitor>(&mut self, v: &V)

pub fn mut_accept_mut<V: MutVisitorMut>(&mut self, v: &mut V)