p4::lexer

Enum Kind

pub enum Kind {

    Const,
    Header,
    Typedef,
    Control,
    Struct,
    Action,
    Parser,
    Table,
    Size,
    Key,
    Exact,
    Ternary,
    Lpm,
    Range,
    Actions,
    DefaultAction,
    Entries,
    In,
    InOut,
    Out,
    Transition,
    State,
    Select,
    Apply,
    Package,
    Extern,
    If,
    Else,
    Return,
    Bool,
    Error,
    Bit,
    Varbit,
    Int,
    String,
    AngleOpen,
    AngleClose,
    CurlyOpen,
    CurlyClose,
    ParenOpen,
    ParenClose,
    SquareOpen,
    SquareClose,
    Semicolon,
    Comma,
    Colon,
    Underscore,
    PoundInclude,
    PoundDefine,
    Backslash,
    Forwardslash,
    DoubleEquals,
    NotEquals,
    Equals,
    Plus,
    Minus,
    Mod,
    Dot,
    Mask,
    LogicalAnd,
    And,
    Bang,
    Tilde,
    Shl,
    Pipe,
    Carat,
    GreaterThanEquals,
    LessThanEquals,
    IntLiteral(i128),
    Identifier(String),
    BitLiteral(u16, u128),
    SignedLiteral(u16, i128),
    TrueLiteral,
    FalseLiteral,
    StringLiteral(String),
    Eof,

}

Variants

Const

Header

Typedef

Control

Struct

Action

Parser

Table

Size

Key

Exact

Ternary

Lpm

Range

Actions

DefaultAction

Entries

In

InOut

Out

Transition

State

Select

Apply

Package

Extern

If

Else

Return

Bool

Error

Bit

Varbit

Int

String

AngleOpen

AngleClose

CurlyOpen

CurlyClose

ParenOpen

ParenClose

SquareOpen

SquareClose

Semicolon

Comma

Colon

Underscore

PoundInclude

PoundDefine

Backslash

Forwardslash

DoubleEquals

NotEquals

Equals

Plus

Minus

Mod

Dot

Mask

LogicalAnd

And

Bang

Tilde

Shl

Pipe

Carat

GreaterThanEquals

LessThanEquals

IntLiteral(i128)

An integer literal. The following are literal examples and their associated types. - 10 : int - 8s10 : int<8> - 2s3 : int<2> - 1s1 : int<1>

Identifier(String)

BitLiteral(u16, u128)

A bit literal. The following a literal examples and their associated types. - 8w10 : bit<8> - 1w1 : bit<1> First element is number of bits (prefix before w) second element is value (suffix after w).

SignedLiteral(u16, i128)

A signed literal. The following a literal examples and their associated types. - 8s10 : bit<8> - 1s1 : bit<1> First element is number of bits (prefix before w) second element is value (suffix after w).

TrueLiteral

FalseLiteral

StringLiteral(String)

Eof

End of file.